PR-Chatbot/pyproject.toml
[project]
name = "pr-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "flask>=3.1.1",
    "githubkit>=0.12.14",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-google-genai>=2.1.5",
    "langchain-openai>=0.3.25",
    "langgraph>=0.4.9",
    "nltk>=3.9.1",
    "numpy>=2.3.1",
    "python-dotenv>=1.1.1",
    "rich>=14.0.0",
    "tqdm>=4.67.1",
    "transformers>=4.52.4",
    "unidiff>=0.7.5",
    "uvicorn>=0.35.0",
]
